plat/kvm: support PKE bit on x86
Provided that PKU is advertised by cpuid (EAX=7, ECX=0, EXC bit 3),
Intel Memory Protection Keys can be enabled by setting bit 22 (PKE) in
CR4.

Since we do not want to pay the cost of MPK-related code when MPK is
not enabled, we introduce a new HAVE_X86PKU property; MPK code is
compile-guarded so that it does not get compiled in without
HAVE_X86PKU. HAVE_X86PKU will be set by PKU kernel libraries later on.

At boot time, if HAVE_X86PKU is enabled, the kernel now checks whether
PKU is advertised by cpuid. If not, it aborts the boot process. The
underlying idea is that images compiled with HAVE_X86PKU are
*specialized* to be executed on PKU-enabled hardware. Hence, provided
that HAVE_X86PKU is set, there is a guarantee that PKU is supported
*and* enabled, effectively removing the need for checks later at
runtime. This might benefit performance when the domain switching rate
is high.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@manchester.ac.uk>